Bounding Boxes and Frustum Culling Added BoundingBoxes for Meshes, calculated during the import of mdoels Added BoundingOrientedBoxes for Objects, both world and local aligned Added BoundingFrustum to Cameras Added ShouldDraw function to Objects which takes in the Camera's frustum and checks it against the world aligned OBB ShouldDraw is checked when drawing Objects, Knits and shadows but not sprites Setting the view for a camera now decomposes the matrix and sets the position and rotation Thetis now can set the view for a camera

+++ Achilles/shaders/PPGammaCorrection.cpp
@@ -41,7 +41,7 @@ void PPGammaCorrection::ApplyGammaCorrection(std::shared_ptr<CommandList> comman
     uint32_t height = (uint32_t)h;
 
     GammaCorrectionCB0 gammaCorrectionCB0;
-    gammaCorrectionCB0.RcpBufferDim = Vector2(1 / width, 1 / height);
+    gammaCorrectionCB0.RcpBufferDim = Vector2(1 / (float)width, 1 / (float)height);
     gammaCorrectionCB0.GammaCorrection = GammaCorrection;
 
     commandList->SetPipelineState(gammaCorrectionShader->pipelineState);
